requirements.txt
appdirs==1.4.4
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
colorlog==6.8.2
Django==5.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Faker==27.0.0
gitdb==4.0.11
GitPython==3.1.43
idna==3.7
octo-framework==1.0.1
pillow==10.4.0
python-dateutil==2.9.0.post0
requests==2.32.3
smmap==5.0.1
sqlparse==0.5.1
urllib3==2.2.2
